+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - Level completion as a goal.
 - Complete all the action stages enabled to unlock the Perfect Chaos fight.
+- Added option to auto-start missions.
 
 ### Changed